[merge]

+++ src/pages/mypage/UserPage.jsx
@@ -52,9 +52,16 @@ const UserPage = () => {
   };
 
   const userDelete = async () => {
+    const deletePassword = prompt("회원탈퇴를 위해 비밀번호를 입력하세요.");
+
+    if (!deletePassword) {
+      alert("비밀번호를 입력하세요 :(");
+      return;
+    }
+    console.log("이게 뭔데? : ", deletePassword);
     try {
       const res = await axios.delete("/api/user", {
-        params: { userId: userData.userId },
+        Data: { userId: userData.userId, upw: deletePassword },
       });
 
       if (
